NextJS examples: Move non-route JS files outside of `app/`
