+++ examples/nextjs-flow/lib/transcript-manager.ts
