+++ examples/nextjs-real-time-transcription/components/AudioVisualizer.tsx
