+++ examples/nextjs-real-time-transcription/components/MicrophoneSelect.tsx
