+++ examples/nextjs-real-time-transcription/components/Output.tsx
