+++ examples/nextjs-real-time-transcription/components/Status.tsx
