+++ examples/nextjs-real-time-transcription/lib/constants.ts
